LPS-76521 Move constants to internal
