+++ modules/apps/foundation/portal-security-sso/portal-security-sso-token/src/main/java/com/liferay/portal/security/sso/token/internal/constants/TokenConfigurationKeys.java
@@ -12,14 +12,19 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.token.constants;
+package com.liferay.portal.security.sso.token.internal.constants;
 
 /**
- * @author Michael C. Han
+ * @author Mika Koivisto
  */
-public class TokenConstants {
+public class TokenConfigurationKeys {
 
-	public static final String SERVICE_NAME =
-		"com.liferay.portal.security.sso.token";
+	public static final String AUTH_ENABLED = "enabled";
+
+	public static final String IMPORT_FROM_LDAP = "importFromLDAP";
+
+	public static final String LOGOUT_REDIRECT_URL = "logoutRedirectURL";
+
+	public static final String USER_HEADER = "userTokenName";
 
 }
